requirements.txt
anthropic>=0.40.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
